+++ benches/benchmarks.rs
@@ -13,6 +13,7 @@ fn end_to_end_benchmark() {
         log_level: logger::LogLevel::Debug,
         markup_types: vec![MarkupType::Markdown],
         ignore_links: vec![],
+        match_file_extension: false,
         ignore_path: vec![],
         root_dir: None,
         no_web_links: false,