+++ src/link_validator/file_system.rs
@@ -6,30 +6,8 @@ use async_std::path::PathBuf;
 use std::path::MAIN_SEPARATOR;
 use walkdir::WalkDir;
 
-pub async fn check_filesystem(source: &str, target: &str, config: &Config) -> LinkCheckResult {
-    let mut normalized_link = target
-        .replace('/', &MAIN_SEPARATOR.to_string())
-        .replace('\\', &MAIN_SEPARATOR.to_string());
-    if let Some(idx) = normalized_link.find('#') {
-        warn!(
-            "Strip everything after #. The chapter part ´{}´ is not checked.",
-            &normalized_link[idx..]
-        );
-        normalized_link = normalized_link[..idx].to_string();
-    }
-    let mut fs_link_target = Path::new(&normalized_link).to_path_buf();
-    if normalized_link.starts_with(MAIN_SEPARATOR) && config.root_dir.is_some() {
-        match canonicalize(&config.root_dir.as_ref().unwrap()).await {
-            Ok(new_root) => fs_link_target = new_root.join(Path::new(&normalized_link[1..])),
-            Err(e) => panic!(
-                "Root path could not be converted to an absolute path. Does the directory exit? {}",
-                e
-            ),
-        }
-    }
-
-    debug!("Check file system link target {:?}", target);
-    let target = absolute_target_path(source, &fs_link_target).await;
+pub async fn check_filesystem(target: &str, config: &Config) -> LinkCheckResult {
+    let target = Path::new(target);
     debug!("Absolute target path {:?}", target);
     let error_result = LinkCheckResult::Failed("Target path not found.".to_string());
     if target.exists().await {
@@ -76,6 +54,36 @@ pub async fn check_filesystem(source: &str, target: &str, config: &Config) -> Li
     } else {
         error_result
     }
+}
+
+pub async fn resolve_target_link(source: &str, target: &str, config: &Config) -> String {
+    let mut normalized_link = target
+        .replace('/', &MAIN_SEPARATOR.to_string())
+        .replace('\\', &MAIN_SEPARATOR.to_string());
+    if let Some(idx) = normalized_link.find('#') {
+        warn!(
+            "Strip everything after #. The chapter part ´{}´ is not checked.",
+            &normalized_link[idx..]
+        );
+        normalized_link = normalized_link[..idx].to_string();
+    }
+    let mut fs_link_target = Path::new(&normalized_link).to_path_buf();
+    if normalized_link.starts_with(MAIN_SEPARATOR) && config.root_dir.is_some() {
+        match canonicalize(&config.root_dir.as_ref().unwrap()).await {
+            Ok(new_root) => fs_link_target = new_root.join(Path::new(&normalized_link[1..])),
+            Err(e) => panic!(
+                "Root path could not be converted to an absolute path. Does the directory exit? {}",
+                e
+            ),
+        }
+    }
+
+    debug!("Check file system link target {:?}", target);
+    absolute_target_path(source, &fs_link_target)
+        .await
+        .to_str()
+        .expect("Could not resolve target path")
+        .to_string()
 }
 
 async fn absolute_target_path(source: &str, target: &PathBuf) -> PathBuf {