+++ src/link_validator/mod.rs
@@ -1,14 +1,17 @@
 mod file_system;
 mod http;
-mod link_type;
 mod mail;
 
+pub mod link_type;
+
+use crate::link_extractors::link_extractor::MarkupLink;
 use crate::link_validator::file_system::check_filesystem;
 use crate::link_validator::http::check_http;
 use crate::Config;
-use link_type::get_link_type;
-use link_type::LinkType;
 use mail::check_mail;
+
+pub use link_type::get_link_type;
+pub use link_type::LinkType;
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum LinkCheckResult {
@@ -19,35 +22,42 @@ pub enum LinkCheckResult {
     NotImplemented(String),
 }
 
-pub async fn check(link_source: &str, link_target: &str, config: &Config) -> LinkCheckResult {
-    info!("Check link {} => {}.", &link_source, &link_target);
+pub async fn resolve_target_link(
+    link: &MarkupLink,
+    link_type: &LinkType,
+    config: &Config,
+) -> String {
+    if link_type == &LinkType::FileSystem {
+        file_system::resolve_target_link(&link.source, &link.target, &config).await
+    } else {
+        link.target.to_string()
+    }
+}
+
+pub async fn check(link_target: &str, link_type: &LinkType, config: &Config) -> LinkCheckResult {
+    info!("Check link {}.", &link_target);
     if config.ignore_links.iter().any(|m| m.is_match(link_target)) {
         return LinkCheckResult::Ignored("Ignore link because of ignore-links option.".to_string());
     }
-    match get_link_type(link_target) {
-        None => {
-            LinkCheckResult::Failed(format!("Could not determine link type of {}.", link_target))
-        }
-        Some(link_type) => match link_type {
-            LinkType::FTP => LinkCheckResult::NotImplemented(format!(
-                "Link type '{:?}' is not supported yet...",
-                &link_type
-            )),
-            LinkType::UnknownUrlSchema => LinkCheckResult::NotImplemented(
-                "Url schema is unknown and cannot be checked.".to_string(),
-            ),
-            LinkType::Mail => check_mail(link_target),
-            LinkType::HTTP => {
-                if config.no_web_links {
-                    LinkCheckResult::Ignored(
-                        "Ignore web link because of the no-web-link flag.".to_string(),
-                    )
-                } else {
-                    check_http(link_target).await
-                }
+    match link_type {
+        LinkType::FTP => LinkCheckResult::NotImplemented(format!(
+            "Link type '{:?}' is not supported yet...",
+            &link_target
+        )),
+        LinkType::UnknownUrlSchema => LinkCheckResult::NotImplemented(
+            "Link type is not implemented yet and cannot be checked.".to_string(),
+        ),
+        LinkType::Mail => check_mail(link_target),
+        LinkType::HTTP => {
+            if config.no_web_links {
+                LinkCheckResult::Ignored(
+                    "Ignore web link because of the no-web-link flag.".to_string(),
+                )
+            } else {
+                check_http(link_target).await
             }
-            LinkType::FileSystem => check_filesystem(link_source, &link_target, &config).await,
-        },
+        }
+        LinkType::FileSystem => check_filesystem(&link_target, &config).await,
     }
 }
 
@@ -61,8 +71,8 @@ mod tests {
         let mut config = Config::default();
         config.ignore_links = vec![wildmatch::WildMatch::new("http?*")];
         let result = check(
-            "NotImportant",
             "https://doesNotExist.me/even/less/likelly",
+            &LinkType::HTTP,
             &config,
         )
         .await;